requirements.txt
# Core ML/DL
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0

# Data & utilities
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0

# Image processing
opencv-python>=4.8.0
Pillow>=9.5.0
albumentations>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Metaheuristics / optimization
deap>=1.4.1
# pygmo>=2.19.0 

# Explainability
shap>=0.44.0
grad-cam>=1.5.0

# App / reports
streamlit>=1.27.0
jupyter>=1.0.0
notebook>=7.0.0

# Database (optional)
sqlalchemy>=2.0.0
